+++ projects/frontend/es-app/src/Upgrade.cpp
@@ -107,7 +107,7 @@ void Upgrade::Run()
 void Upgrade::ReceiveSyncMessage()
 {
   // Volatile popup
-  mWindow.InfoPopupAdd(new GuiInfoPopup(mWindow, mPopupMessage, PatronInfo::Instance().IsPatron() ? 10 : 30, PopupType::Recalbox));
+  mWindow.InfoPopupAdd(new GuiInfoPopup(mWindow, mPopupMessage, 10, PopupType::Recalbox));
 
   // Messagebox
   if (!mMessageBoxMessage.empty())